Roll PowerApi generators into Server generators
This cleanup is a first step towards making it possible to configure whether
or not to generate Power API's on a service-by-service basis rather than
globally
