+++ codegen/src/main/scala/akka/grpc/gen/scaladsl/ScalaServerCodeGenerator.scala
@@ -6,13 +6,13 @@ package akka.grpc.gen.scaladsl
 
 import akka.grpc.gen.Logger
 import com.google.protobuf.compiler.PluginProtos.CodeGeneratorResponse
-import templates.ScalaServer.txt._
+import templates.ScalaServer.txt.{ Handler, PowerApiTrait }
 
 case class ScalaServerCodeGenerator(powerApis: Boolean = false) extends ScalaCodeGenerator {
   override def name = "akka-grpc-scaladsl-server"
 
   override def perServiceContent = super.perServiceContent + generateHandler() ++ (
-    if (powerApis) Set(generateHandler(powerApis))
+    if (powerApis) Set(generateHandler(powerApis), generatePowerApiTrait)
     else Set.empty
   )
 
@@ -21,6 +21,14 @@ case class ScalaServerCodeGenerator(powerApis: Boolean = false) extends ScalaCod
     b.setContent(Handler(service, powerApis).body)
     b.setName(s"${service.packageDir}/${service.name}${if (powerApis) "PowerApi" else ""}Handler.scala")
     logger.info(s"Generating Akka gRPC service${if (powerApis) " power API" else ""} handler for ${service.packageName}.${service.name}")
+    b.build
+  }
+
+  val generatePowerApiTrait: (Logger, Service) => CodeGeneratorResponse.File = (logger, service) => {
+    val b = CodeGeneratorResponse.File.newBuilder()
+    b.setContent(PowerApiTrait(service).body)
+    b.setName(s"${service.packageDir}/${service.name}PowerApi.scala")
+    logger.info(s"Generating Akka gRPC service power API interface for ${service.packageName}.${service.name}")
     b.build
   }
 }
